FIX recontruction foulders
