+++ src/App.jsx
@@ -8,6 +8,65 @@ import Events from "./pages/Events/Events";
 import NotFound from "./pages/NotFound/NotFound";
 import Profile from "./pages/Profile/Profile";
 
+const users =  [
+  {
+    name:"Adam",
+    role:"admin",
+    email:"fgfdgdfg@gmail.com",
+    number:"1234567890",
+    password:"1111",
+    organization:"eLogic"
+  },
+  {
+    name:"Yeva",
+    role:"manager",
+    email:"sdfsdg@gmail.com",
+    number:"1234567890",
+    password:"1111",
+    organization:"SoftServe"
+  },
+  {
+    name:"Tanya",
+    role:"manager",
+    email:"fsdfae@gmail.com",
+    number:"1234567890",
+    password:"1111",
+    organization:"InventorSoft"
+  },
+  {
+    name:"Matroskin",
+    role:"manager",
+    email:"retdsf@gmail.com",
+    number:"1234567890",
+    password:"1111",
+    organization:"ASD"
+  },
+  {
+    name:"Leopold",
+    role:"user",
+    email:"sdfgag@gmail.com",
+    number:"1234567890",
+    password:"1111",
+    organization:"SharpMinds"
+  },
+  {
+    name:"Tiger",
+    role:"user",
+    email:"sdfsx@gmail.com",
+    number:"1234567890",
+    password:"1111",
+    organization:"EPAM"
+  },
+  {
+    name:"Tom",
+    role:"user",
+    email:"sdffx@gmail.com",
+    number:"1234567890",
+    password:"1111",
+    organization:"Pub24"
+  }
+]
+
 const App = () => {
   return (
     <>
