--- src/pages/Managers/Managers.scss
@@ -1,9 +1,0 @@
-.managers{
-  min-height: 79vh;
-  &-img{
-    display: block;
-    background: #282c34;
-    width: 7vw;
-    height: 7vw;
-  }
-}
