+++ src/pages/Managers/components/ManagersList/ManagersList.scss
@@ -1,0 +1,9 @@
+.managers{
+  min-height: 79vh;
+  &-img{
+    display: block;
+    background: #282c34;
+    width: 7vw;
+    height: 7vw;
+  }
+}
