--- src/pages/Members/Members.scss
@@ -1,9 +1,0 @@
-.members{
-  min-height: 79vh;
-  &-img{
-    display: block;
-    background: #282c34;
-    width: 7vw;
-    height: 7vw;
-  }
-}
