+++ src/pages/Members/components/MembersList/MembersList.scss
@@ -1,0 +1,9 @@
+.members{
+  min-height: 79vh;
+  &-img{
+    display: block;
+    background: #282c34;
+    width: 7vw;
+    height: 7vw;
+  }
+}
